Updates
Learning Requirements IMplement, Loot System Revamped, Better Drop Probablities.
Other cleanups and some template changes.

+++ Darkages.Server/Systems/Loot/Interfaces/IWeighable.cs
@@ -8,6 +8,6 @@
         /// <summary>
         /// The weight of this item.
         /// </summary>
-        int Weight { get; set; }
+        double Weight { get; set; }
     }
 }
